+++ changelog.md
@@ -5,3 +5,9 @@
 ### v0.1.1
 - November 2, 2024
 - Revise ReadME.md
+
+### v0.1.2
+- November 4, 2024
+- It finally passes the data using function
+- Revise the HTML
+- Merged all the Components to render at once